Enable ability to update target_url for exisitng entries in .flyrc
Prior to this commit if an entry was already present in the flyrc then it would use the target_url from flyrc - there was no way to update without doing it manually.
Reworked the logic around targets in the flyrc so that if a concourse_url or ca_cert is specified then it will overwrite the existing entry in flyrc and reset the auth token back to nothing.