Handle hijack -u use case.
The hijack use case is unique because it doessn't require a target to be passed.
It will instead search the flyrc to find a matching API url and use that as the target.
This caused complications where duplicate API urls are in use by different targets becase we need to ensure that the auth token has been updated for all targets where the API URL is the same.